Move SQLTableStore hooks into their own file
